Dont include current site in available sites

+++ includes/classes/InternalConnections/NetworkSiteConnection.php
@@ -452,10 +452,10 @@ class NetworkSiteConnection extends Connection {
 		$sites            = get_sites();
 		$authorized_sites = array();
 
-		$current_blog_id = get_current_blog_id();
+		$current_blog_id = (int) get_current_blog_id();
 
 		foreach ( $sites as $site ) {
-			$blog_id = $site->blog_id;
+			$blog_id = (int) $site->blog_id;
 
 			if ( $blog_id === $current_blog_id ) {
 				continue;
